requirements.txt
Django==4.0.1
djangorestframework==3.13.1
django-heroku==0.3.1
whitenoise==5.3.0
psycopg2==2.9.3
python-dotenv==0.19.2
django-cors-headers==3.11.0
drf-yasg==1.20.0
django-extensions==3.1.5
channels==3.0.4
channels_redis
asgiref~=3.5.0
channels-redis==3.3.1
daphne==3.0.2
schedule==1.1.0
Celery==5.2.3
